requirements.txt
aiogram==3.15.0
aiosqlite==0.20.0
python-dotenv==1.0.1
google-generativeai==0.8.3
apscheduler==3.10.4
